requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
spacy==3.7.0
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.2.0
anthropic==0.6.0
pytest==7.4.2
httpx==0.25.0
sqlalchemy==2.0.21
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
prometheus-client==0.17.1
loguru==0.7.2 